rpm: Fix installation path of SysV startup script
Do not use /etc/init.d for SysV init scripts for RPM-based
linux distros.  The directory should be /etc/rc.d/init.d.

When using /etc/init.d and no chkconfig or initscripts packages
are installed first, this package was creating /etc/init.d as
a directory.  Unfortunately, if you wanted to later install
a package that brought in the chkconfig package, the RPM
transaction would fail because it assumes it can create
a symlink in /etc/init.d when a directory already exists.

A bug was filed for this here:
https://github.com/osquery/osquery/issues/7949